# Repair fast stream for bootstrap and decommission