--- a/dht/range_streamer.hh
+++ b/dht/range_streamer.hh
@@ -148,7 +148,7 @@ private:
         return *_token_metadata_ptr;
     }
 public:
-    future<> stream_async();
+    future<> stream_async(std::function<void(size_t)> progress_updater = nullptr);
     size_t nr_ranges_to_stream();
 private:
     distributed<replica::database>& _db;