repair: Introduce fast_stream for bootstrap and decommission with RBNO

When the peer node to sync data with is only one in rbno, we could use
classic streaming instad of repair to speed up the bootstrap process.

The speed up comes mostly from the fact that classic streaming could
batch the ranges and stream multiple ranges in a stream session, i.e.,
10% of total ranges per stream session. When the number of the ranges is
high, this batch helps a lot, especially in large clusters where the
number of ranges is big and the data in each range is relatively small.

Before:
using rbno, 5m partitions, 88 seconds
```
[asias@hjpc2 scylla]$ cat mycluster/node3.log.txt|grep bootstrap_with_repair|grep keyspace=ks
repair - bootstrap_with_repair: started with keyspace=ks201, nr_ranges=273, nr_tables=1
repair - bootstrap_with_repair: finished with keyspace=ks201, nr_ranges=273, nr_tables=1, duration=88.698s
```

using classic stream, 5m partitions, 45 seconds
```
range_streamer - Bootstrap with [ea424ee8-6059-4159-a9a2-997451c96276, f34358f9-17d8-4380-9282-778568142281] for keyspace=ks201 started, nodes_to_stream=2
range_streamer - Bootstrap with ea424ee8-6059-4159-a9a2-997451c96276 for keyspace=ks201, streaming [0, 13) out of 134 ranges
stream_session - [Stream #ee3f2270-8f86-11f0-a904-ec12ac9fa53d] Executing streaming plan for Bootstrap-ks201-index-0 with peers={ea424ee8-6059-4159-a9a2-997451c96276}, master
range_streamer - Bootstrap with f34358f9-17d8-4380-9282-778568142281 for keyspace=ks201, streaming [0, 13) out of 138 ranges
stream_session - [Stream #ee3f2271-8f86-11f0-a904-ec12ac9fa53d] Executing streaming plan for Bootstrap-ks201-index-0 with peers={f34358f9-17d8-4380-9282-778568142281}, master
...
range_streamer - Bootstrap with f34358f9-17d8-4380-9282-778568142281 for keyspace=ks201 succeeded, took 40.509224 seconds
range_streamer - Bootstrap with ea424ee8-6059-4159-a9a2-997451c96276 for keyspace=ks201 succeeded, took 44.787827 seconds
```

After:
using rbno + fast stream optimization, 5m partitions, 45 seconds
```
[asias@hjpc2 scylla]$ cat mycluster/node3.log.txt|grep bootstrap_with_repair|grep keyspace=ks
repair - bootstrap_with_repair: started with keyspace=ks201, nr_ranges=273, nr_tables=1
repair - bootstrap_with_repair: finished with keyspace=ks201, nr_ranges=273, nr_tables=1, duration=45.1912s
```

With this patch, bootstrap with RBNO enabled and bootstrap with RNBO
disabled have the same level of performance.

Closes #24664


performance optimization, no backport