@denesb 2025-09-17
> Closes #24664

Please use `Fixes`. Although `Closes` is also a valid keyword, we de-facto use it to close the PR with the merge commit, with `Fixes` being distinctly used to close a fixed issue.