@denesb 2025-09-17
> performance optimization, no backport

We may want to backport this actually, @mykaul what do you think?