@avikivity 2025-09-17
Oh, it's for normal operation when the peer in the same rack is there and we don't need true repair.