@asias 2025-09-19
> @asias how does RBNO (and this new version) interact with `tombstone-gc: {'mode': 'repair'}`?

Repair within RBNO does not update the repair history. So it does matter if streaming or bootstrap is used for the sake of tombstone gc mode.