@asias 2025-09-24
> This PR simply uses non-RBNO streaming for bootrstapping a new node and decommissioning. Why do we need this (IMO unnecessary) PR which tries to hide the fact instead of simply explicitly NOT using RBNO for these operations by changing the default of `allowed_repair_based_node_ops` as suggested in #24664?
> 
> The latter is a trivial change which achieves the same result exactly with **much** less code required. Another advantage of #24664 is that it's explicit and self-explanatory, while this proposal is exactly the opposite.
> 
> @denesb I'm strongly against this change.

- You are welcome to suggest what you want to achieve in the end from the support point of view. When it can be achieved in multiple ways, it does not have to be what you suggested exactly.  

- How data is transferred internally inside repair is an implementation detail.  Nothing prevents repair to use direct stream if it is possible.  

- The change in this PR can benefit all node operations in addition to bootstrap and decommission. For example, for replace or rebuild operations, when the peer node is one, the direct stream optimization can turned on automatically.

- This PR can be extended to use direct stream when no data is present on local node and use repair if data is present to resume.  This keeps both the benefits of fast stream and the ability to resume for bootstrap and decommission. 

- This PR helps to unify the bootstrap code path. Instead of having two places to figure out what ranges to fetch and from which node to fetch.  

- The code in this PR is straightforward, it uses the existing streamer infrastructure. 

So, I think this is a better solution than what you proposed.