@bhalevy 2025-09-25
> I support this PR, I think having a single way to do node ops is where we should be heading, even if under the hood it uses multiple ways to effectively sync the data. We should eventually retire the old streaming code (just the session management, but not the actual data sync), drop `enable_repair_based_node_ops` and `allowed_repair_based_node_ops` and make `rbno` the only backend for node ops (with a choice of data sync method maybe).

@denesb how about backportability?
Isn't changing the defaults for `allowed_repair_based_node_ops` to remove bootstrap and decommission safer than backporting this PR, for released versions?