@denesb 2025-09-25
> > I support this PR, I think having a single way to do node ops is where we should be heading, even if under the hood it uses multiple ways to effectively sync the data. We should eventually retire the old streaming code (just the session management, but not the actual data sync), drop `enable_repair_based_node_ops` and `allowed_repair_based_node_ops` and make `rbno` the only backend for node ops (with a choice of data sync method maybe).
> 
> @denesb how about backportability? Isn't changing the defaults for `allowed_repair_based_node_ops` to remove bootstrap and decommission safer than backporting this PR, for released versions?

Yes, for sure changing the default is safer, for older releases.
@asias please send a PR to change the default as @vladzcloudius suggested, this will be backported.
This PR will restore the earlier defaults and enable the fast streaming for boostrap and decomission. This way <2025.3 can keep using non-RBNO for bootstrap/decomission and 2025.4+ will use the new fast-streaming with RBNO.