@vladzcloudius 2025-09-25
> I support this PR, I think having a single way to do node ops is where we should be heading, even if under the hood it uses multiple ways to effectively sync the data. We should eventually retire the old streaming code (just the session management, but not the actual data sync), drop `enable_repair_based_node_ops` and `allowed_repair_based_node_ops` and make `rbno` the only backend for node ops (with a choice of data sync method maybe).

While I'm now not against this PR in general I disagree that we should ever hide the fact that there are actually two very different ways of streaming data between nodes and we should always have ways to allow explicit enforcing one or another.

 We should also be very explicit both in our documentation and in our defaults which type of streaming is advised and in which operations.

And this PR by no means addresses this requirement.