@vladzcloudius 2025-09-25
> You are welcome to suggest what you want to achieve in the end from the support point of view. When it can be achieved in multiple ways, it does not have to be what you suggested exactly.

@asias this PR doesn't achieve what I have requested and I've explained it in https://github.com/scylladb/scylladb/issues/24664.

It is orthogonal at best and is potentially harmful since it creates a potential for a regression.